Mention Plack & frameworks under web programming.
For someone looking for information on web programming with Perl, we should
really be showing them more modern, nice ways, not encouraging them to write CGI
scripts.

I'd like to put a list of links to some of the more modern Perl frameworks here,
but I could be considered biased towards one of them, so I'll leave this for
others to decide whether it's a good idea.