change versioning algorithm to 5.YYYYmmdd
It was also considered a possibility to use the version of the next perl
release as our version, but that would likely lead to difficulty in
coordinating releases, and makes it more difficult to have more than one
release in the same month.  Also this method has been used by
Module::CoreList for a few years and has worked well there.